Separated Program and Function views better from other UI logic
Also fixed a bug with recycling views
